requirements.txt
asgiref==3.8.1
django==5.1.2
django-jazzmin==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.1
psycopg2-binary
pytz==2024.2
PyYAML==6.0.2
ruff==0.7.1
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
python-dotenv